feat(package): removed the input argument and better project structure
This a major update that adds a new name and changes the CLI usage.

BREAKING CHANGE:
`-i` (`--input`) is removed as an option. Also added a command for `images`
which was default. This is for future support.

The command is also renamed to `dte` _(devtools-timeline-export). For a
reason being, it can export more than images. Old command name is kept
for backwards compatibility. Related to #3

Closes #6
