+++ package.json
@@ -18,10 +18,13 @@
   ],
   "homepage": "https://github.com/ilicmarko/devtools-timeline-images#readme",
   "bin": {
-    "devtools-timeline-images": "lib/index.js"
+    "devtools-timeline-images": "lib/index.js",
+    "dte": "lib/index.js"
   },
   "dependencies": {
     "chalk": "^2.4.2",
+    "fs-extra": "^8.0.0",
+    "inquirer": "^6.3.1",
     "yargs": "^13.2.2"
   }
 }
